Rename `uuid` to `identifier` to prevent conflicts
Some development environments get confused when standard library
name is used for custom module, let's make it easier for them.

Fix #1267.
