+++ mypy.ini
@@ -9,6 +9,7 @@ non_interactive = True
 
 files = tmt/__init__.py,
         tmt/__main__.py,
+        tmt/identifier.py,
         tmt/plugins/__init__.py,
         tmt/steps/__init__.py,
         tmt/steps/prepare/ansible.py,
@@ -16,8 +17,7 @@ files = tmt/__init__.py,
         tmt/steps/prepare/shell.py,
         tmt/steps/provision/artemis.py,
         tmt/templates.py,
-        tmt/utils.py,
-        tmt/uuid.py
+        tmt/utils.py
 
 # This is probably temporary: don't follow imports to other tmt packages
 # and modules, but once the list of files is complete, this should be dropped
